requirements.txt
# Thumbnail Generator MVP - Dependencies

# Frontend Framework
streamlit>=1.32.0

# Core Image Processing
Pillow>=10.2.0
numpy>=1.26.0

# Gradio API Client
gradio-client>=0.15.0
requests>=2.31.0

# Type Hints
typing-extensions>=4.9.0

# Utilities
python-dotenv>=1.0.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Performance
numba==0.58.1

# Optional: Advanced Image Processing
opencv-python==4.8.1.78
scikit-image==0.22.0